将时间获取从UTC时间改为BootTime时间

+++ services/network_search/src/network_search_handler.cpp
@@ -34,7 +34,7 @@ std::condition_variable NetworkSearchManager::cv_;
 static const int32_t REQ_INTERVAL = 30;
 const int32_t SATELLITE_STATUS_ON = 1;
 const size_t MCC_LEN = 3;
-constexpr int32_t SECOND_TO_MILLI = 1000;
+constexpr int64_t SECOND_TO_MILLI = 1000;
 const std::string PERMISSION_PUBLISH_SYSTEM_EVENT = "ohos.permission.PUBLISH_SYSTEM_COMMON_EVENT";
 const std::map<uint32_t, NetworkSearchHandler::NsHandlerFunc> NetworkSearchHandler::memberFuncMap_ = {
     { RadioEvent::RADIO_SIM_STATE_CHANGE,
